Reorganized BisecationMethod and new main.cpp
The compiler is having trouble running so will need to fix that
(might have something to do with the launcher or where main.cpp
is located). While writing out a basic test in main noticed that the
endpoints could be located anywhere as long as they have seperate signs,
so changed the naming and made them be private vars for the
BisectionMethod object.