Updated so that compiler now works
Had to change the tasks.json file to compile and link all .cpp files not
just the ones in the base directory. tasks.json was also updated to
allow for c++11 extension.
Additionally added beginning functionality to allow for user inputted
functions however the majority is hard coded without importing a
library (will save for when working on the GUI).